feat(): 增加签到功能
增加签到功能; 修改网关中积分扣减逻辑，同步扣减，异步退还

+++ glory-api-common/src/main/java/com/hry/gloryapi/common/model/entity/InterfaceInfo.java
@@ -84,7 +84,7 @@ public class InterfaceInfo implements Serializable {
     /**
      * 积分
      */
-    private Integer integral;
+    private Long integral;
 
     /**
      * 创建人
